Refactor: List 컴포넌트 수정 * delete 이벤트 연결 * click 이벤트 연결
