+++ Frontend/tarzan/src/components/bookmark/HouseTap.vue
@@ -15,35 +15,45 @@
         <p>집 추가하기</p>
       </div>
       <div class="house-list-wrapper">
-        <SwipeItem
-          v-for="(bookmark, index) in bookmarks"
-          :key="index"
-          @delete="handleDelete(index)"
-          @click="navigateToDetailPage(bookmark)"
-        >
-          <div class="house-item-wrapper">
-            <HouseItem :house="bookmark" />
-          </div>
-        </SwipeItem>
+        <List
+          :fetchItems="fetchBookmarks"
+          :params="params"
+          :onDelete="handleDelete"
+          :onClick="handleCLick">
+          <template #item="{ item, index, onClick, onDelete }">
+            <SwipeItem
+              :key="item.bookmarkIdx"
+              @click="onClick(item)"
+              @delete="onDelete(item, index)">
+              <div class="house-item-wrapper" style="background-color: aqua">
+                <BookmarkItem :bookmark="item" />
+              </div>
+            </SwipeItem>
+          </template>
+        </List>
       </div>
-
-      <!-- <NonContent v-if="list.length === 0" :value="'북마크한 집이 존재하지 않습니다.'" /> -->
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
 /**  */
-import { ref, onMounted, defineProps, defineExpose, defineEmits, watch } from "vue";
+import {
+  ref,
+  onMounted,
+  defineProps,
+  defineExpose,
+  defineEmits,
+  watch,
+} from "vue";
 
 /** component load */
 import AddressHouseSearch from "@/components/common/AddressHouseSearch.vue";
 
 import houseAddImg from "@/assets/icons/Plus/Pluse.png";
-import HouseAddSrc from "@/assets/icons/Plus/Style=Outlined.svg";
-import NonContent from "@/components/common/NonContent.vue";
-import HouseItem from "@/components/bookmark/HouseItem.vue";
+import BookmarkItem from "@/components/bookmark/BookmarkItem.vue";
 import TagButtonGroup from "@/components/common/TagButtonGroup.vue";
+import List from "@/components/common/List.vue";
 import SwipeItem from "@/components/common/SwipeItem.vue";
 
 /** data, plugin load */
@@ -66,22 +76,9 @@ function navigateToMap() {
   router.push("/bookmark/map");
 }
 
-const navigateToDetailPage = (house) => {
+const handleDelete = async (bookmark: any) => {
   try {
-    if (!house.bookmarkIdx) throw new Error("bookmarkIdx 없음");
-    router.push({ name: "BookMarkDetail", params: { id: house.bookmarkIdx } });
-  } catch (e) {
-    console.error("navigate 에러:", e);
-  }
-};
-
-
-const handleDelete = async (idx) => {
-  console.log("삭제");
-  console.log(bookmarks.value[idx]);
-  const bookmarkIdx = bookmarks.value[idx].bookmarkIdx;
-  try {
-    bookmarks.value.splice(idx, 1);
+    const bookmarkIdx = bookmark.bookmark_id;
     const response = await axiosInstance.delete(`/v1/bookmark/${bookmarkIdx}`);
   } catch {
     console.error("잘못된 요청입니다.");
@@ -98,38 +95,15 @@ const closeAddressSearch = () => {
   showAddressSearch.value = false;
 };
 
-const list = ref([]);
-list.value = [
-  {
-    bookmarkIdx: 1,
-    house_name: "집 이름 1",
-    house_address: "주소 1",
-    house_category: "카테고리 1",
-    house_review_score: 4.5,
-    house_latitude: 37.5665,
-    house_longitude: 126.978,
-    created_at: "2025.03.16 10:00:00",
-  },
-  {
-    bookmarkIdx: 2,
-    house_name: "집 이름 2",
-    house_address: "주소 2",
-    house_category: "카테고리 2",
-    house_review_score: 4.5,
-    house_latitude: 37.5665,
-    house_longitude: 126.978,
-    created_at: "2025.03.16 10:00:00",
-  },
-];
-
 const handleCLick = (bookmark) => {
+  console.log(bookmark);
   router.push({
     name: "BookMarkDetail",
-    params: { id: bookmark.bookmarkIdx },
+    params: { id: bookmark.bookmark_id },
     query: { bookmarkObj: JSON.stringify(bookmark) },
   });
 };
-
+/** tag button 관련 */
 const tagOptions = [
   { label: "전체", value: "ALL" },
   { label: "점검 전", value: "CHECK_PENDING" },
@@ -138,46 +112,37 @@ const tagOptions = [
 
 const selectedTag = ref("ALL"); // 단일 선택용
 
-const bookmarks = ref<House[]>([]);
+/** bookmark item 관련 */
+watch(selectedTag, (newSeletedTag) => {
+  params.value.status = newSeletedTag;
+});
 
-const fetchRecentHouses = async () => {
+const params = ref({
+  size: 10,
+  status: "ALL", // 나중에 필터링 추가
+  sortBy: "최신순",
+});
+
+const fetchBookmarks = async (page: number, params: any) => {
   try {
     const response = await axiosInstance.get(`/v1/bookmark`, {
       params: {
-        size: 3,
-        page: 0,
-        sortBy: "최신순",
-        status: selectedTag.value,
+        ...params,
+        page: page,
       },
     });
 
     if (response.data.success) {
-      console.log(response.data);
-      bookmarks.value = response.data.data.list
-        .map((item) => ({
-          bookmarkIdx: item.bookmark_id,
-          house_name: item.bookmark_house_name || "이름 없음",
-          house_address: item.bookmark_house_address,
-          house_category: item.bookmark_house_category || "카테고리 없음",
-          house_latitude: item.bookmark_house_latitude,
-          house_longitude: item.bookmark_house_longitude,
-          created_at: item.bookmark_created_at, // created_at도 같이 저장
-        }))
-        .filter((house) => house.bookmarkIdx !== undefined); // bookmarkIdx가 undefined인 항목은 제거
-
-      console.log(bookmarks.value);
+      return response.data.data.list;
     } else {
       console.error("Failed to fetch data:", response.data.message);
+      return [];
     }
   } catch (error) {
     console.error("API request error:", error);
+    return [];
   }
 };
-
-onMounted(fetchRecentHouses);
-watch(selectedTag, (newSeletedTag) => {
-  fetchRecentHouses();
-});
 </script>
 
 <style lang="scss" scoped>
